SSOR-100 refactor and update tests
